+++ lib/querly/preprocessor.rb
@@ -19,8 +19,30 @@ module Querly
     end
 
     def run!(source_code)
-      output, status = Open3.capture2({ 'RUBYOPT' => nil }, command, stdin_data: source_code)
-      raise Error.new(status: status, command: command) unless status.success?
+      stdin_read, stdin_write = IO.pipe
+      stdout_read, stdout_write = IO.pipe
+
+      writer = Thread.new do
+        stdin_write.print source_code
+        stdin_write.close
+      end
+
+      output = ""
+
+      reader = Thread.new do
+        while (line = stdout_read.gets)
+          output << line
+        end
+      end
+
+      succeeded = system(command, in: stdin_read, out: stdout_write)
+      stdout_write.close
+
+      writer.join
+      reader.join
+
+      raise Error.new(status: $?, command: command) unless succeeded
+
       output
     end
   end