+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## master
 
 * Make `[conditional]` be aware of safe-navigation-operator (@pocke) #30
+* Make preprocessors be aware of `bundle exec`.
+  When `querly` is invoked with `bundle exec`, so are preprocessors, and vice vesa.
  
 ## 0.7.0 (2017-08-22)